requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2

# Visualization (optional but okay)
matplotlib==3.8.4
seaborn==0.13.2

# Machine Learning
scikit-learn==1.3.2
imbalanced-learn==0.11.0
xgboost==2.0.3
joblib==1.4.2
mlflow
# API
fastapi==0.110.0
uvicorn==0.29.0

# Kafka
kafka-python==2.0.2

# Database
mysql-connector-python==8.3.0

# HTTP
requests==2.31.0

# Env
python-dotenv==1.0.1